+++ data/shader/vulkan/tile_border.vert
@@ -3,7 +3,7 @@
 
 layout (location = 0) in vec2 inVertex;
 #ifdef TW_TILE_TEXTURED
-layout (location = 1) in uvec4 inVertexTexCoord;
+layout (location = 1) in vec4 inVertexTexCoord;
 #endif
 
 layout(push_constant) uniform SPosBO {
